update docs and organize tests
