+++ tests_playwright/README.md
@@ -2,8 +2,12 @@
 
 Playwright tests live in this directory.
 
-- Current setup, commands, fixture architecture, suite status, and troubleshooting:
-  `docs/PLAYWRIGHT_GUIDE.md`
+- Setup and running tests:
+  `docs/PLAYWRIGHT_SETUP_AND_RUN.md`
+- Test structure and fixture architecture:
+  `docs/PLAYWRIGHT_ARCHITECTURE.md`
+- Writing and extending tests:
+  `docs/PLAYWRIGHT_WRITING_TESTS.md`
 - Historical background, migration notes, and prior failure modes:
   `docs/PLAYWRIGHT_HISTORY.md`
