+++ tests_playwright/project/table_react/rename_table_column.spec.ts
@@ -3,8 +3,8 @@ import {
     createTemporaryProjectViaSyntheticAnndata,
     type SyntheticAnndataTemporaryProjectHandle,
     waitForProjectReady,
-} from "../utils/projectFixtures";
-import { saveCurrentView } from "../utils/helpers";
+} from "../../utils/projectFixtures";
+import { saveCurrentView } from "../../utils/helpers";
 
 const TABLE_CHART_TITLE = "Rename Column Table";
