+++ tests_playwright/project/table_react/soft_delete_column.spec.ts
@@ -3,7 +3,7 @@ import {
     createTemporaryProjectViaSyntheticAnndata,
     type TemporaryProjectHandle,
     waitForProjectReady,
-} from "../utils/projectFixtures";
+} from "../../utils/projectFixtures";
 
 type ProjectSetupInfo = {
     cloneSourceName: string;
